+++ src/main/java/uk/gov/pay/connector/app/ConnectorModule.java
@@ -24,12 +24,12 @@ import uk.gov.pay.connector.common.validator.RequestValidator;
 import uk.gov.pay.connector.gateway.GatewayClient;
 import uk.gov.pay.connector.gateway.GatewayClientFactory;
 import uk.gov.pay.connector.gateway.PaymentProviders;
-import uk.gov.pay.connector.gateway.model.request.records.CardAuthoriseRequestFactory;
-import uk.gov.pay.connector.gateway.model.request.records.WorldpayCardAuthoriseRequestFactory;
-import uk.gov.pay.connector.gateway.model.request.records.WorldpayMotoAuthoriseRequestFactory;
+import uk.gov.pay.connector.gateway.model.CardAuthoriseRequestFactory;
 import uk.gov.pay.connector.gateway.stripe.StripeSdkClientFactory;
 import uk.gov.pay.connector.gateway.stripe.StripeSdkWrapper;
 import uk.gov.pay.connector.gateway.templates.WorldpayRequestTemplateBuilder;
+import uk.gov.pay.connector.gateway.worldpay.WorldpayCardAuthoriseRequestFactory;
+import uk.gov.pay.connector.gateway.worldpay.WorldpayMotoAuthoriseRequestFactory;
 import uk.gov.pay.connector.gatewayaccount.resource.GatewayAccountRequestValidator;
 import uk.gov.pay.connector.gatewayaccount.service.GatewayAccountServicesFactory;
 import uk.gov.pay.connector.paymentprocessor.service.CardExecutorService;
