+++ src/main/java/uk/gov/pay/connector/gateway/adyen/utils/AuthoriseRequestPayloadToGatewayOrderConverter.java
@@ -1,11 +1,8 @@
-package uk.gov.pay.connector.gateway.adyen;
+package uk.gov.pay.connector.gateway.adyen.utils;
 
 import jakarta.ws.rs.core.MediaType;
-import uk.gov.pay.connector.app.adyen.AdyenGatewayConfig;
 import uk.gov.pay.connector.gateway.GatewayOrder;
 import uk.gov.pay.connector.gateway.adyen.request.json.AuthoriseRequestPayload;
-import uk.gov.pay.connector.gateway.model.request.AuthorisationGatewayRequest;
-import uk.gov.pay.connector.gateway.model.request.RecurringPaymentAuthorisationGatewayRequest;
 import uk.gov.pay.connector.util.JsonObjectMapper;
 
 import static uk.gov.pay.connector.gateway.model.OrderRequestType.AUTHORISE;
