+++ src/main/java/uk/gov/pay/connector/gateway/worldpay/WorldpayCardAuthoriseRequestFactory.java
@@ -1,7 +1,8 @@
-package uk.gov.pay.connector.gateway.model.request.records;
+package uk.gov.pay.connector.gateway.worldpay;
 
 import jakarta.inject.Inject;
 import uk.gov.pay.connector.gateway.model.request.CardAuthorisationGatewayRequest;
+import uk.gov.pay.connector.gateway.model.request.records.WorldpayCardAuthoriseRequest;
 import uk.gov.service.payments.commons.model.AuthorisationMode;
 
 import java.util.Optional;
