+++ src/test/java/uk/gov/pay/connector/gateway/adyen/utils/AdyenMerchantAccountHelperTest.java
@@ -1,4 +1,4 @@
-package uk.gov.pay.connector.gateway.model.request.records;
+package uk.gov.pay.connector.gateway.adyen.utils;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.extension.ExtendWith;
