+++ src/test/java/uk/gov/pay/connector/gateway/stripe/util/StripeAuthorisationRejectedCodeMapperTest.java
@@ -1,6 +1,7 @@
-package uk.gov.pay.connector.gateway.model;
+package uk.gov.pay.connector.gateway.stripe.util;
 
 import org.junit.jupiter.api.Test;
+import uk.gov.pay.connector.gateway.model.MappedAuthorisationRejectedReason;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.is;
