+++ src/test/java/uk/gov/pay/connector/gateway/util/ChargeFrontendUrlHelperTest.java
@@ -1,4 +1,4 @@
-package uk.gov.pay.connector.gateway.model.request.records;
+package uk.gov.pay.connector.gateway.util;
 
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
